+++ sdk/core/iOS/MoEReactBridge/MoEngageReactPluginInfo.h
@@ -1,2 +1,3 @@
 // Generated file, do not edit
 #define MOE_REACT_PLUGIN_VERSION @"9.0.0"
+#define MOE_REACT_PLUGIN_NAME @"react_native"