+++ sdk/expo/CHANGELOG.md
@@ -6,6 +6,8 @@
 - [NA] Fixing the failing cases due to types/node version update
 - Android
   - [minor] Compatibility changes for "Bundling `MoEFireBaseMessagingService` in Plugin's manifest to reduce integration steps for developers using FCM Push." in core plugin
+- iOS
+  - [minor] Added no-code SDK file based initialization
 
 # 25-08-2025